+++ components/style/properties/longhand/border.mako.rs
@@ -3,7 +3,6 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 <%namespace name="helpers" file="/helpers.mako.rs" />
-
 <% from data import Method %>
 
 <% data.new_style_struct("Border", inherited=False, gecko_ffi_name="nsStyleBorder",
@@ -62,3 +61,5 @@
                               "computed::BorderRadiusSize::zero()",
                               "parse")}
 % endfor
+
+${helpers.single_keyword("box-decoration-break", "slice clone", products="gecko")}
