+++ components/style/properties/longhand/outline.mako.rs
@@ -3,7 +3,6 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 <%namespace name="helpers" file="/helpers.mako.rs" />
-
 <% from data import Method %>
 
 <% data.new_style_struct("Outline",
